JBIDE-22168: Make sure tempFile is deleted on Windows as well

+++ plugins/org.jboss.tools.hibernate.runtime.v_5_1/src/org/jboss/tools/hibernate/runtime/v_5_1/internal/ConfigurationFacadeImpl.java
@@ -76,7 +76,7 @@ public class ConfigurationFacadeImpl extends AbstractConfigurationFacade {
 		File tempFile = null;
 		IConfiguration result = null;
 		try {
-			tempFile = File.createTempFile(document.toString(), "cfg.xml");
+			tempFile = File.createTempFile(document.toString() + ".cfg", "xml");
 			DOMSource domSource = new DOMSource(document);
 			StringWriter stringWriter = new StringWriter();
 			StreamResult stream = new StreamResult(stringWriter);
@@ -89,8 +89,12 @@ public class ConfigurationFacadeImpl extends AbstractConfigurationFacade {
 			result = configure(tempFile);
 		} catch(IOException | TransformerException e) {
 			throw new RuntimeException("Problem while configuring", e);
-		} finally {
-			tempFile.delete();
+		}
+		System.out.println("*********** Checking for tempFile: " + tempFile.getAbsolutePath() + " ***************************");
+		if (tempFile.exists()) {
+			System.out.println("*********** Deleting tempFile: " + tempFile.getAbsolutePath() + " ***************************");
+			boolean success = tempFile.delete();
+			System.out.println("*********** ===> " + success + "*************");
 		}
 		return result;
 	}
